Modify client.c to print Fibonacci numbers
Originally, client.c just prints the Fibonacci numbers in hexadecimal
, but it is added with some functions about big number now.

So, there has a function called bign_to_str(), which converts a big
number with 128 bits to a string so that the string can be a readable
number for human (in decimal).

Also, client.c is not only to get the Fibonacci numbers but also
calculate and obtain the three different execution time, such as user
space execution, kernel space and kernel-to-user execution time.

Finally, the three time will be recored in the file 'time.csv'.

Then, here adds a gnuplot script to plot a graph about the trend
of the three different time. To illustrate the increasing of the
time when the program wants to get the larger Fibonacci numbers.